requirements.txt
agate==1.6.3
aiobotocore==2.1.2
aiohttp==3.8.1
aioitertools==0.10.0
aiosignal==1.2.0
apturl==0.5.2
asn1crypto==0.24.0
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
botocore==1.23.24
Brlapi==0.6.6
cachetools==5.0.0
certifi==2018.1.18
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.4
cloudpickle==2.0.0
colorama==0.4.4
command-not-found==0.3
coverage==4.0.3
crcmod==1.7
croniter==1.3.4
cryptography==2.1.4
cupshelpers==1.0
dask==2022.2.1
dbt-bigquery==1.0.0
dbt-core==1.0.3
dbt-extractor==0.4.0
defer==1.0.6
distlib==0.3.4
distributed==2022.2.1
distro-info==0.18ubuntu0.18.04.1
Django==2.1.2
docker==5.0.3
ecdsa==0.17.0
enum34==1.1.10
filelock==3.6.0
frozenlist==1.3.0
fsspec==2022.2.0
future==0.18.2
google-api-core==2.6.0
google-auth==2.6.0
google-cloud-aiplatform==1.11.0
google-cloud-bigquery==2.34.2
google-cloud-core==2.2.2
google-cloud-secret-manager==2.9.2
google-cloud-storage==2.2.1
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.55.0
grpc-google-iam-v1==0.12.3
grpcio==1.44.0
grpcio-status==1.44.0
HeapDict==1.0.1
hologram==0.0.14
httplib2==0.11.3
idna==2.6
importlib-metadata==4.11.2
importlib-resources==5.4.0
iniconfig==1.1.1
intuit-oauth==1.1.1
isodate==0.6.1
Jinja2==2.11.3
jmespath==0.10.0
jsonschema==3.1.1
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
leather==0.3.4
locket==0.2.1
Logbook==1.5.3
loguru==0.6.0
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==2.0.1
marshmallow==3.15.0
marshmallow-oneofschema==3.0.1
mashumaro==2.9
minimal-snowplow-tracker==0.0.2
mock==4.0.3
msgpack==1.0.3
multidict==6.0.2
mypy-extensions==0.4.3
netifaces==0.10.4
networkx==2.7.1
numpy==1.22.3
oauth==1.0.1
oauthlib==3.2.0
olefile==0.45.1
packaging==21.3
pandas==1.4.1
parsedatetime==2.4
partd==1.2.0
pendulum==2.1.2
pexpect==4.2.1
Pillow==5.1.0
pip==22.0.3
platformdirs==2.5.1
pluggy==1.0.0
prefect==1.1.0
proto-plus==1.20.3
protobuf==3.19.4
psutil==5.9.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.21
pycrypto==2.6.1
pycups==1.9.73
pygit2==1.9.0
PyGObject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==3.0.7
pyRFC3339==1.0
pyrsistent==0.18.1
pytest==7.1.0
pytest-cov==2.5.0
python-apt==1.6.5+ubuntu0.7
python-box==6.0.1
python-coveralls==2.9.3
python-dateutil==2.8.2
python-debian==0.1.32
python-dotenv==0.19.2
python-jose==3.3.0
python-slugify==6.1.1
pytimeparse==1.1.8
pytz==2021.3
pytzdata==2020.1
pyxdg==0.25
PyYAML==6.0
reportlab==3.4.0
requests==2.27.1
requests-oauthlib==1.3.1
requests-unixsocket==0.1.5
rsa==4.8
s3fs==2022.2.0
SecretStorage==2.3.1
setuptools==60.9.3
simplejson==3.13.2
six==1.11.0
sortedcontainers==2.4.0
sqlparse==0.4.2
system-service==0.3
systemd-python==234
tabulate==0.8.9
tblib==1.7.0
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
tornado==6.1
typing_extensions==4.1.1
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.9
usb-creator==0.3.3
virtualenv==20.13.2
wadllib==1.3.2
websocket-client==1.3.1
Werkzeug==2.0.3
wheel==0.37.1
wrapt==1.14.0
xkit==0.0.0
yarl==1.7.2
zict==2.1.0
zipp==3.7.0
zope.interface==4.3.2
